huffmantree微调代码

+++ Tree/HuffmanTree/src/main.cpp
@@ -8,7 +8,7 @@ int main() {
 	double w[] = { 2,4,6,1,10 };
 	int n = sizeof(w) / sizeof(w[0]);
 
-	HuffmanTree<int> htree(w, n);
+	HuffmanTree<double> htree(w, n);
 
 	htree.showTree();
